# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API of software to describe impressions of books.",
        "title": "book_Impressions_back",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "host": "Secret",
    "basePath": "/",
    "paths": {
        "/api/account/login": {
            "post": {
                "description": "JWT certification",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "If the infomation passed in the request body matches the data in the table, a cookie will be issued.",
                "responses": {
                    "200": {
                        "description": "200 OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "400 incrrect password",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "401 unauthenticated",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "404 Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/account/logout": {
            "post": {
                "description": "JWT certification",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "If the cookie exists, delete it.",
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "401": {
                        "description": "401 unauthenticated",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/account/nowuser": {
            "get": {
                "description": "Browse Account table.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "Show infomation about the currently logged in user.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Account"
                        }
                    },
                    "401": {
                        "description": "401 unauthenticated",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Browse Account table.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "Updates information about the currently logged in user.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Account"
                        }
                    },
                    "401": {
                        "description": "401 unauthenticated",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "404 Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Browse Account table.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "Deletes infomation about the currently logged in user.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Account"
                        }
                    },
                    "401": {
                        "description": "401 unauthenticated",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "404 Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/account/signup": {
            "post": {
                "description": "Use the account table.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "Register Account infomation in the database.",
                "responses": {
                    "201": {
                        "description": "201 Created",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "409 It is already registered",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/impression/:id": {
            "get": {
                "description": "Can be executed only at login.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "impression"
                ],
                "summary": "Display of impressions.",
                "responses": {
                    "200": {
                        "description": "200 OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "401 unauthenticated",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "404 Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Can be executed only at login.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "impression"
                ],
                "summary": "Update of impressions.",
                "responses": {
                    "200": {
                        "description": "200 OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "401 unauthenticated",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "404 Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Can be executed only at login.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "impression"
                ],
                "summary": "Delete impressions.",
                "responses": {
                    "200": {
                        "description": "200 OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "401 unauthenticated",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "404 Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/impressions": {
            "get": {
                "description": "Can be executed only at login.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "impressions"
                ],
                "summary": "List of impressions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Impression"
                        }
                    },
                    "401": {
                        "description": "401 unauthenticated",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Can be executed only at login.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "impressions"
                ],
                "summary": "impression registration process.",
                "responses": {
                    "200": {
                        "description": "200 OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "401 unauthenticated",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/impressions/search/:bookid": {
            "get": {
                "description": "Can be executed only at login.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "impressions"
                ],
                "summary": "Search for books using Bookid as a key.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Impression"
                        }
                    },
                    "401": {
                        "description": "401 unauthenticated",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "404 Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Account": {
            "type": "object",
            "properties": {
                "create_at": {
                    "type": "string"
                },
                "delete_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "uid": {
                    "type": "string"
                },
                "update_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "models.Impression": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "bookid": {
                    "type": "string"
                },
                "booktitle": {
                    "type": "string"
                },
                "create_at": {
                    "type": "string"
                },
                "delete_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "imageurl": {
                    "type": "string"
                },
                "isbn10": {
                    "type": "string"
                },
                "isbn13": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "uid": {
                    "type": "string"
                },
                "update_at": {
                    "type": "string"
                }
            }
        }
    }
}